+++ src/runtime/core/rpc/local_rpc_backend.cc
@@ -18,10 +18,15 @@ struct convert<aimrt::runtime::core::rpc::LocalRpcBackend::Options> {
   static Node encode(const Options& rhs) {
     Node node;
 
+    node["timeout_executor"] = rhs.timeout_executor;
+
     return node;
   }
 
   static bool decode(const Node& node, Options& rhs) {
+    if (node["timeout_executor"])
+      rhs.timeout_executor = node["timeout_executor"].as<std::string>();
+
     return true;
   }
 };
@@ -39,6 +44,29 @@ void LocalRpcBackend::Initialize(YAML::Node options_node,
     options_ = options_node.as<Options>();
 
   rpc_registry_ptr_ = rpc_registry_ptr;
+
+  if (!options_.timeout_executor.empty()) {
+    AIMRT_CHECK_ERROR_THROW(
+        get_executor_func_,
+        "Get executor function is not set before initialize.");
+
+    auto timeout_executor = get_executor_func_(options_.timeout_executor);
+
+    AIMRT_CHECK_ERROR_THROW(
+        timeout_executor,
+        "Get timeout executor '{}' failed.", options_.timeout_executor);
+
+    client_tool_.RegisterTimeoutExecutor(timeout_executor);
+    client_tool_.RegisterTimeoutHandle(
+        [](std::shared_ptr<runtime::core::rpc::ClientInvokeWrapper>&& client_invoke_wrapper_ptr) {
+          client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_TIMEOUT);
+        });
+
+    AIMRT_TRACE("Local rpc backend enable the timeout function, use '{}' as timeout executor.",
+                options_.timeout_executor);
+  } else {
+    AIMRT_TRACE("Local rpc backend does not enable the timeout function.");
+  }
 
   options_node = options_;
 }
@@ -86,10 +114,23 @@ bool LocalRpcBackend::TryInvoke(
     const std::shared_ptr<ClientInvokeWrapper>& client_invoke_wrapper_ptr) noexcept {
   assert(state_.load() == State::Start);
 
+  namespace util = aimrt::common::util;
+
   std::string_view pkg_path = client_invoke_wrapper_ptr->pkg_path;
   std::string_view module_name = client_invoke_wrapper_ptr->module_name;
   std::string_view func_name = client_invoke_wrapper_ptr->func_name;
 
+  // 检查ctx，当前只支持一个server url，且在plugin层配置
+  std::string_view to_addr =
+      client_invoke_wrapper_ptr->ctx_ref.GetMetaValue(AIMRT_RPC_CONTEXT_KEY_TO_ADDR);
+
+  if (!to_addr.empty()) {
+    auto pos = to_addr.find("://");
+    if (pos == std::string_view::npos) return false;
+    if (to_addr.substr(0, pos) != Name()) return false;
+  }
+
+  // 需要由本后端处理。此行之后只能使用callback报错，不能返回false
   const auto& service_func_wrapper_map = rpc_registry_ptr_->GetServiceFuncWrapperMap();
   const auto& client_func_wrapper_map = rpc_registry_ptr_->GetClientFuncWrapperMap();
 
@@ -103,13 +144,12 @@ bool LocalRpcBackend::TryInvoke(
   std::string_view service_pkg_path, service_module_name;
 
   // url: local://rpc/func_name?pkg_path=xxxx&module_name=yyyy
-  std::string_view to_addr = client_invoke_wrapper_ptr->ctx_ref.GetToAddr();
   if (!to_addr.empty()) {
-    auto url = aimrt::common::util::ParseUrl<std::string_view>(to_addr);
+    auto url = util::ParseUrl<std::string_view>(to_addr);
     if (url) {
       assert(url->protocol == Name());
-      service_pkg_path = aimrt::common::util::GetValueFromStrKV(url->query, "pkg_path");
-      service_module_name = aimrt::common::util::GetValueFromStrKV(url->query, "module_name");
+      service_pkg_path = util::GetValueFromStrKV(url->query, "pkg_path");
+      service_module_name = util::GetValueFromStrKV(url->query, "module_name");
     }
   }
 
@@ -198,9 +238,8 @@ bool LocalRpcBackend::TryInvoke(
     return true;
   }
 
-  // 不在一个pkg内，需要经过序列化
-  std::string serialization_type(
-      client_invoke_wrapper_ptr->ctx_ref.GetSerializationType());
+  // 不在一个pkg内，需要经过序列化，并启用timeout功能
+  std::string serialization_type(client_invoke_wrapper_ptr->ctx_ref.GetSerializationType());
 
   // 找注册的client方法
   auto get_client_func_wrapper_ptr_func = [&]() -> const ClientFuncWrapper* {
@@ -217,21 +256,35 @@ bool LocalRpcBackend::TryInvoke(
   };
   const auto* client_func_wrapper_ptr = get_client_func_wrapper_ptr_func();
 
+  // 将回调等内容记录下来
+  uint32_t cur_req_id = req_id_++;
+  auto record_ptr = client_invoke_wrapper_ptr;
+  bool ret = client_tool_.Record(
+      cur_req_id,
+      client_invoke_wrapper_ptr->ctx_ref.Timeout(),
+      std::move(record_ptr));
+
+  if (!ret) [[unlikely]] {
+    // 一般不太可能出现
+    AIMRT_WARN("Failed to record msg.");
+    client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_CLI_UNKNOWN);
+    return true;
+  }
+
+  // client req序列化
   aimrt::util::BufferArray buffer_array;
   auto client_req_type_support_ref = aimrt::util::TypeSupportRef(client_func_wrapper_ptr->req_type_support);
 
-  // client req序列化
   bool serialize_ret = client_req_type_support_ref.Serialize(
       serialization_type, client_invoke_wrapper_ptr->req_ptr, buffer_array.AllocatorNativeHandle(), buffer_array.BufferArrayNativeHandle());
 
   if (!serialize_ret) {
     // 序列化失败
     AIMRT_ERROR(
-        "Req serialization failed in local rpc backend, serialization_type {}, "
-        "pkg_path: {}, module_name: {}, func_name: {}",
+        "Req serialization failed in local rpc backend, serialization_type {}, pkg_path: {}, module_name: {}, func_name: {}",
         serialization_type, pkg_path, module_name, func_name);
 
-    client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_CLI_SERIALIZATION_FAILDE);
+    client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_CLI_SERIALIZATION_FAILED);
 
     return true;
   }
@@ -247,11 +300,10 @@ bool LocalRpcBackend::TryInvoke(
   if (!deserialize_ret) {
     // 反序列化失败
     AIMRT_FATAL(
-        "Rsp deserialization failed in local rpc backend, serialization_type {}, "
-        "pkg_path: {}, module_name: {}, func_name: {}",
+        "Rsp deserialization failed in local rpc backend, serialization_type {}, pkg_path: {}, module_name: {}, func_name: {}",
         serialization_type, pkg_path, module_name, func_name);
 
-    client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_CLI_DESERIALIZATION_FAILDE);
+    client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_CLI_DESERIALIZATION_FAILED);
 
     return true;
   }
@@ -266,10 +318,19 @@ bool LocalRpcBackend::TryInvoke(
       [this,
        service_func_wrapper_ptr,
        client_func_wrapper_ptr,
-       client_invoke_wrapper_ptr,
+       cur_req_id,
        service_req_ptr,
        service_rsp_ptr,
        serialization_type{std::move(serialization_type)}](uint32_t code) {
+        auto msg_recorder = client_tool_.GetRecord(cur_req_id);
+        if (!msg_recorder) [[unlikely]] {
+          // 未找到记录，说明此次调用已经超时了，走了超时处理后删掉了记录
+          AIMRT_TRACE("Can not get req id {} from recorder.", cur_req_id);
+          return;
+        }
+
+        auto client_invoke_wrapper_ptr = std::move(*msg_recorder);
+
         aimrt::util::BufferArray buffer_array;
 
         auto service_rsp_type_support_ref = aimrt::util::TypeSupportRef(service_func_wrapper_ptr->rsp_type_support);
@@ -281,13 +342,12 @@ bool LocalRpcBackend::TryInvoke(
         if (!serialize_ret) {
           // 序列化失败
           AIMRT_ERROR(
-              "Rsp serialization failed in local rpc backend, serialization_type {}, "
-              "pkg_path: {}, module_name: {}, func_name: {}",
+              "Rsp serialization failed in local rpc backend, serialization_type {}, pkg_path: {}, module_name: {}, func_name: {}",
               serialization_type, service_func_wrapper_ptr->pkg_path,
               service_func_wrapper_ptr->module_name,
               service_func_wrapper_ptr->func_name);
 
-          client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_SVR_SERIALIZATION_FAILDE);
+          client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_SVR_SERIALIZATION_FAILED);
 
           return;
         }
@@ -296,18 +356,19 @@ bool LocalRpcBackend::TryInvoke(
 
         // client rsp 反序列化
         bool deserialize_ret = client_rsp_type_support_ref.Deserialize(
-            serialization_type, *TO_AIMRT_BUFFER_ARRAY_VIEW(buffer_array.BufferArrayNativeHandle()), client_invoke_wrapper_ptr->rsp_ptr);
+            serialization_type,
+            *TO_AIMRT_BUFFER_ARRAY_VIEW(buffer_array.BufferArrayNativeHandle()),
+            client_invoke_wrapper_ptr->rsp_ptr);
 
         if (!deserialize_ret) {
           // 反序列化失败
-          AIMRT_FATAL(
-              "Req deserialization failed in local rpc backend, serialization_type {}, "
-              "pkg_path: {}, module_name: {}, func_name: {}",
+          AIMRT_ERROR(
+              "Req deserialization failed in local rpc backend, serialization_type {}, pkg_path: {}, module_name: {}, func_name: {}",
               serialization_type, service_func_wrapper_ptr->pkg_path,
               service_func_wrapper_ptr->module_name,
               service_func_wrapper_ptr->func_name);
 
-          client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_SVR_DESERIALIZATION_FAILDE);
+          client_invoke_wrapper_ptr->callback(AIMRT_RPC_STATUS_SVR_DESERIALIZATION_FAILED);
 
           return;
         }
@@ -320,6 +381,14 @@ bool LocalRpcBackend::TryInvoke(
       service_rsp_ptr.get(), service_callback.NativeHandle());
 
   return true;
+}
+
+void LocalRpcBackend::RegisterGetExecutorFunc(
+    const std::function<aimrt::executor::ExecutorRef(std::string_view)>& get_executor_func) {
+  AIMRT_CHECK_ERROR_THROW(
+      state_.load() == State::PreInit,
+      "Function can only be called when state is 'PreInit'.");
+  get_executor_func_ = get_executor_func;
 }
 
 }  // namespace aimrt::runtime::core::rpc