+++ document/release_notes.md
@@ -28,6 +28,8 @@
 - 【非兼容性修改】去除channel的context manager，现在可以直接new一个context；
 - 【非兼容性修改】去除rpc的context manager，现在可以直接new一个context；
 - 【非兼容性修改】原来的协程RPC Proxy/Service接口重命名为CoProxy/CoService，不在支持老命名方式；
+- 优化了rpc status Tostring方法的输出；
+- local rpc backend 支持timeout功能；
 - 新增 log_control_plugin：
   - 提供了运行时查看、修改日志等级的接口；
   - 添加了相关示例；